% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5shiny.R
\docType{package}
\name{lab5shiny}
\alias{lab5shiny}
\alias{lab5shiny-package}
\title{Visual Interaction Tool For Swedish Election Results}
\description{
Implements an interactive tool for the Swedish election results where you can
see the results for each muniplicity. The default select is Stockholms Kommun.
Uses the elect_viz package
}
