% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5shiny.R
\name{server}
\alias{server}
\title{The server component for the shiny application. Doing the backend of the visualization.}
\usage{
server(input, output)
}
\arguments{
\item{input}{Takes a Shiny input source}

\item{output}{Takes a shiny output source}
}
\description{
The server component for the shiny application. Doing the backend of the visualization.
}
