% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Server.R
\name{shinyServer}
\alias{shinyServer}
\title{shinyServer}
\arguments{
\item{function(input,}{output) : an anonymous function taking count of what
is chosen by the user and gives the corresponding output}
}
\description{
depending of the selected län and party, an histogram is
       given with the score of this party in the different kommuner of
       this län.
}
\details{

}
\author{
Niclas Lovsjö & Maxime Bonneau
}

