% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5shiny.R
\docType{data}
\name{ui}
\alias{ui}
\title{The UI for the shiny application}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\value{
a fluidPage object
}
\description{
The UI for the shiny application
}
\keyword{datasets}
