% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5shiny.R
\docType{data}
\name{viz}
\alias{viz}
\title{The RC class to interface with the data}
\format{An object of class \code{elect_viz} of length 1.}
\usage{
viz
}
\value{
RC class object elect_viz
}
\description{
The RC class to interface with the data
}
\keyword{datasets}
